requirements.txt
Django~=4.2.7
psycopg2==2.9.9
Pillow==10.1.0
Faker==20.0.0
requests~=2.31.0
boto3~=1.34.7
django-storages~=1.14.2
django-webpack-loader~=3.0.0